Moved routes vue files to views folder
