+++ package.json
@@ -5,7 +5,7 @@
 	"main": "server.js",
 	"scripts": {
 		"dev": "run-p dev:client dev:server dev:start",
-		"dev:start": "node -e \"setTimeout(()=>{}, 6000)\" && nodemon src/server.js --inspect --ignore",
+		"dev:start": "node -e \"setTimeout(()=>{}, 3000)\" && nodemon src/server.js --inspect --ignore",
 		"dev:client": "webpack-dev-server -d --hot --config config/webpack.client.config.js",
 		"dev:server": "webpack --watch --config config/webpack.server.config.js",
 		"start": "cross-env NODE_ENV=production node src/server.js",
