+++ src/index.template.html
@@ -2,7 +2,7 @@
 <html lang="en">
 
 <head>
-	<title>SSR VUE</title>
+	<title>Van Gogh MMT Tour</title>
 	<meta charset="utf-8">
 </head>
