+++ src/router.js
@@ -1,8 +1,8 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 
-import Home from './components/Home.vue';
-import About from './components/About.vue';
+import Home from './views/Home.vue';
+import About from './views/About.vue';
 
 Vue.use(Router);
